requirements.txt
Flask==2.2.3
Pillow==9.4.0
rembg==2.0.29
